% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_species_helper_functions.R
\name{all_codes_allowed}
\alias{all_codes_allowed}
\title{Check Species Codes for Compliance with a Given Coding}
\usage{
all_codes_allowed(x, coding)
}
\arguments{
\item{x}{Object of one of the fe_species_id classes or a \code{character}
vector}

\item{coding}{Name of the coding against which \code{x} is to be checked. Must
be one of the entries in the column \code{species_coding} in the data file
\code{\link{species_codings}}}
}
\value{
A list with two elements: 1) \code{rqmt_ok} which is \code{TRUE} if
the requirement is fulfilled and \code{FALSE} otherwise; 2)
\code{err_message}, the body of an error message to be used by the calling
function (mentions all problematic elements of x)
}
\description{
Check Species Codes for Compliance with a Given Coding
}
\keyword{internal}
