% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_stand_helpers_validation.R
\name{check_order_of_magnitude}
\alias{check_order_of_magnitude}
\title{Check a Number's Order of Magnitude Against a Requirement}
\usage{
check_order_of_magnitude(x, min_ok, max_ok)
}
\arguments{
\item{x}{number to check}

\item{min_ok}{an integer power of 10, allowed minimum}

\item{max_ok}{an integer power of 10, allowed maximum}
}
\value{
TRUE if a number x's integer power of 10 is inside the range given by
\code{min_ok} and \code{max_ok} (\code{min_ok} and \code{max_ok} included)
}
\description{
Check a Number's Order of Magnitude Against a Requirement
}
\keyword{internal}
