% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_stand_helpers_construction.R
\name{check_trees_orders_of_magnitude}
\alias{check_trees_orders_of_magnitude}
\title{Check the Orders of Magnitude of the Variables Given in a 'trees' data.frame
as Required in an \emph{fe_stand} object}
\usage{
check_trees_orders_of_magnitude(trees)
}
\arguments{
\item{trees}{data.frame (tibble) structured as required for the 'trees'
component in an \emph{fe_stand} object}
}
\value{
The function issues a warning if it suspects that one of the
variables in \code{trees} does not have the right units. This is the side
effect it is actually called for. It always returns \code{TRUE} invisibly.
}
\description{
Check the Orders of Magnitude of the Variables Given in a 'trees' data.frame
as Required in an \emph{fe_stand} object
}
\keyword{internal}
