% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_species_generic_methods.R
\name{fe_species_get_coding}
\alias{fe_species_get_coding}
\title{Get Name of the Coding Belonging to an \strong{fe_species} Object}
\usage{
fe_species_get_coding(x)
}
\arguments{
\item{x}{An object of one of the supported \code{fe_species} classes}
}
\value{
The requested coding name (\code{character})
}
\description{
Get Name of the Coding Belonging to an \strong{fe_species} Object
}
\examples{
spec_ids <- fe_species_ger_nfi_2012(c("10", "10", "30"))
fe_species_get_coding(spec_ids)

}
