% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_species_generic_methods.R
\name{fe_species_get_coding_table}
\alias{fe_species_get_coding_table}
\title{Get the Coding Table of a Supported \strong{fe_species} Coding}
\usage{
fe_species_get_coding_table(coding)
}
\arguments{
\item{coding}{A character string representing one of the supported codings as
can be requested for a given \strong{fe_species} object with
\code{\link{fe_species_get_coding}}, or one of the entries in the column
\code{species_coding} of the data \code{species_codings}}
}
\value{
A \code{data.frame} (\code{tibble}) representing the requested coding
}
\description{
Get the Coding Table of a Supported \strong{fe_species} Coding
}
\examples{
fe_species_get_coding_table("ger_nfi_2012")
fe_species_get_coding_table("bavrn_state")
fe_species_get_coding_table("tum_wwk_short")

}
