% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_species_generic_methods.R
\name{fe_species_is_tree}
\alias{fe_species_is_tree}
\title{Test Which Elements of a Species Code Vector Denote Tree Species}
\usage{
fe_species_is_tree(x)
}
\arguments{
\item{x}{An object of one of the supported \code{fe_species} classes}
}
\value{
A \code{logical} vector of the same length as \code{x}
}
\description{
Returns a logical vector indicating, for each element of an \strong{fe_species}
vector, whether its code denotes a tree species (\code{TRUE}) or a non-tree
category such as a shrub (\code{FALSE}; see
\code{\link{fe_species_non_tree_codes}}). \code{NA} codes yield \code{NA}.
}
\examples{
spec_ids <- fe_species_bavrn_state(c("10", "60", "80"))
fe_species_is_tree(spec_ids)

}
