% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_species_generic_methods.R
\name{fe_species_non_tree_codes}
\alias{fe_species_non_tree_codes}
\title{Non-Tree Codes of a Species Coding}
\usage{
fe_species_non_tree_codes(coding)
}
\arguments{
\item{coding}{A character string naming one of the supported codings (see
the data \code{\link{species_codings}})}
}
\value{
A \code{character} vector of the coding's non-tree \code{species_id}s
(empty if there are none)
}
\description{
Some species codings contain legal codes that do not correspond to a tree
species and cannot be used in calculations (e.g. a "shrub" category). Such
codes are flagged in the coding table with \code{is_tree = FALSE}. This
function returns them for a given coding. If the coding table has no
\code{is_tree} column (i.e. all of its codes denote tree species), an empty
character vector is returned.
}
\examples{
fe_species_non_tree_codes("bavrn_state")

}
