% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_species_generic_methods.R
\name{format_fe_species}
\alias{format_fe_species}
\title{Generic Formatted Output for \strong{fe_species} Objects}
\usage{
format_fe_species(x, spec_lang = c("code", "eng", "ger", "sci"))
}
\arguments{
\item{x}{Object of one of the supported \code{fe_species_} classes}

\item{spec_lang}{Choice of how species (group) names or id's are displayed.
Supported choices are "code" (default, displays the species codes as they
are), "eng" (English species names), "ger" (German species names), and
"sci" (scientific species names). The names and the codes refer to the
species coding given in the object's attribute \code{species_coding}.}
}
\value{
A \code{character} vector either displaying the original species
codes provided in \code{x}, or the species (group) names in the desired
language
}
\description{
Generic Formatted Output for \strong{fe_species} Objects
}
\keyword{internal}
