% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_q_from_d_q.R
\name{h_q_from_d_q_core}
\alias{h_q_from_d_q_core}
\title{Core Function for Estimating Quadratic Mean Height from Quadratic Mean
Diameter}
\usage{
h_q_from_d_q_core(species_id, d_q_cm)
}
\arguments{
\item{species_id}{when handed to this core function, species_id must be
character and correspond to the tum_wwk_short species coding
\code{\link{fe_species_tum_wwk_short}}}

\item{d_q_cm}{see \code{\link{h_q_from_d_q}}}
}
\value{
see \code{\link{h_q_from_d_q}}
}
\description{
See documentation of \code{\link{h_q_from_d_q}} for most information,
this function is the workhorse behind, and should never be called directly by
a user.
}
\keyword{internal}
