% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_stand_helpers_validation.R
\name{has_no_missing_values}
\alias{has_no_missing_values}
\title{Check if Selected Columns in a Data Frame Contain Missing Values}
\usage{
has_no_missing_values(x, names)
}
\arguments{
\item{x}{a \code{data.frame}}

\item{names}{\code{character} vector containg the names of the columns in
\code{x} to be checked}
}
\value{
A list with two elements: 1) \code{rqmt_ok} which is \code{TRUE} if
the requirement is fulfilled and \code{FALSE} otherwise; 2)
\code{err_message}, the body of an error message to be used by the calling
function (mentions all problematic elements)
}
\description{
Check if Selected Columns in a Data Frame Contain Missing Values
}
\keyword{internal}
