% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_stand_helpers_validation.R
\name{has_required_lengths}
\alias{has_required_lengths}
\title{Check Required Length for a List of Objects}
\usage{
has_required_lengths(x, names_and_rqmts)
}
\arguments{
\item{x}{list of objects to be checked}

\item{names_and_rqmts}{a \code{tibble} that relates object names and the
length they are required to have. Must have two columns, \code{name} (i.e.
the object name, \code{character}), and \code{required} (i.e. the required
length, \code{integer})}
}
\value{
A list with two elements: 1) \code{rqmt_ok} which is \code{TRUE} if
the requirement is fulfilled and \code{FALSE} otherwise; 2)
\code{err_message}, the body of an error message to be used by the calling
function (mentions all problematic elements)
}
\description{
Check Required Length for a List of Objects
}
\keyword{internal}
