% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_stand_helpers_validation.R
\name{has_required_names}
\alias{has_required_names}
\title{Check for Required Elements Existing in an Object}
\usage{
has_required_names(x, required_names)
}
\arguments{
\item{x}{Any object for which \code{names} returns a character vector}

\item{required_names}{Character vector of names required to be in
\code{names(x)}}
}
\value{
A list with two elements: 1) \code{rqmt_ok} which is \code{TRUE} if
the requirement is fulfilled and \code{FALSE} otherwise; 2)
\code{err_message}, the body of an error message to be used by the calling
function (mentions all problematic elements)
}
\description{
Check for Required Elements Existing in an Object
}
\keyword{internal}
