% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_stand_helpers_validation.R
\name{has_required_type_or_class}
\alias{has_required_type_or_class}
\title{Check if an Object has a Required Type or Class}
\usage{
has_required_type_or_class(x, rqmt)
}
\arguments{
\item{x}{object to be checked}

\item{rqmt}{name of the type or class required for \code{x}
(\code{character})}
}
\value{
\code{TRUE} or \code{FALSE}, indicating if the requirement is met or
not
}
\description{
Not a complete coverage of all base types and classes, but tailored to the
needs of \code{\link{validate_fe_stand}}
}
\keyword{internal}
