% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_stand_helpers_validation.R
\name{has_required_types_or_classes}
\alias{has_required_types_or_classes}
\title{Check Required Type or Class for a List of Objects}
\usage{
has_required_types_or_classes(x, names_and_rqmts)
}
\arguments{
\item{x}{a list of objects to be checked}

\item{names_and_rqmts}{a \code{tibble} that relates object names and the
names of the types and classes they are required to have. Must have two
\code{character} columns, \code{name} (i.e. the object name), and
\code{required} (i.e. the name of the required type or class)}
}
\value{
A list with two elements: 1) \code{rqmt_ok} which is \code{TRUE} if
the requirement is met and \code{FALSE} otherwise; 2)
\code{err_message}, the body of an error message to be used by the calling
function (mentions all problematic elements)
}
\description{
Check Required Type or Class for a List of Objects
}
\keyword{internal}
