% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_standard_gnfi3.R
\name{hstd_gnfi3_spec_convert}
\alias{hstd_gnfi3_spec_convert}
\title{Helper function for cascaded species conversion when applying the German
National forest inventory 2012 standard height curve system.
Try to convert given species_id into ger_nfi_2012. If this causes an error
convert to tum_wwk_short instead}
\usage{
hstd_gnfi3_spec_convert(species_id)
}
\description{
Helper function for cascaded species conversion when applying the German
National forest inventory 2012 standard height curve system.
Try to convert given species_id into ger_nfi_2012. If this causes an error
convert to tum_wwk_short instead
}
\keyword{internal}
