% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_stand_helpers_validation.R
\name{is_distinct}
\alias{is_distinct}
\title{Check if Given Variable Combinations in a Data Frame are Distinct}
\usage{
is_distinct(x, combinations)
}
\arguments{
\item{x}{a \code{data.frame}}

\item{combinations}{a \code{list} of \code{character} vectors, where each
describes a variable combination in \code{x} to be checked for being
distinct}
}
\value{
A list with two elements: 1) \code{rqmt_ok} which is \code{TRUE} if
the requirement is fulfilled and \code{FALSE} otherwise; 2)
\code{err_message}, the body of an error message to be used by the calling
function (mentions all problematic elements)
}
\description{
Check if Given Variable Combinations in a Data Frame are Distinct
}
\keyword{internal}
