% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_ccircle_spatial.R
\name{is_fe_ccircle_spatial}
\alias{is_fe_ccircle_spatial}
\title{Check if an Object is an \strong{fe_ccircle_spatial}}
\usage{
is_fe_ccircle_spatial(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{fe_ccircle_spatial}
class
}
\description{
Check if an Object is an \strong{fe_ccircle_spatial}
}
\examples{
strange <- "xyzabc"
is_fe_ccircle_spatial(strange)

}
