% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_ccircle_spatial_notrees.R
\name{is_fe_ccircle_spatial_notrees}
\alias{is_fe_ccircle_spatial_notrees}
\title{Check if an Object is an \strong{fe_ccircle_spatial_notrees} object}
\usage{
is_fe_ccircle_spatial_notrees(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the
\code{fe_ccircle_spatial_notrees} class
}
\description{
Check if an Object is an \strong{fe_ccircle_spatial_notrees} object
}
\examples{
  strange <- "xyzabc"
  is_fe_ccircle_spatial_notrees(strange)

}
