% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_species_bavrn_state.R
\name{is_fe_species_bavrn_state}
\alias{is_fe_species_bavrn_state}
\title{Check if an Object is a \strong{fe_species_bavrn_state} species code vector}
\usage{
is_fe_species_bavrn_state(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the
\code{fe_species_bavrn_state} class
}
\description{
Check if an Object is a \strong{fe_species_bavrn_state} species code vector
}
\examples{
spec_ids <- new_fe_species_bavrn_state(
  as.character(c(40, 40, 30, 30, 60, 60, 60))
)
is_fe_species_bavrn_state(spec_ids)

}
