% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_species_bavrn_state_short.R
\name{is_fe_species_bavrn_state_short}
\alias{is_fe_species_bavrn_state_short}
\title{Check if an Object is a \strong{fe_species_bavrn_state_short} species code vector}
\usage{
is_fe_species_bavrn_state_short(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the
\code{fe_species_bavrn_state_short} class
}
\description{
Check if an Object is a \strong{fe_species_bavrn_state_short} species code vector
}
\examples{
spec_ids <- new_fe_species_bavrn_state_short(
  as.character(1:9)
)
is_fe_species_bavrn_state_short(spec_ids)

}
