% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_species_ger_nfi_2012.R
\name{is_fe_species_ger_nfi_2012}
\alias{is_fe_species_ger_nfi_2012}
\title{Check if an Object is a \strong{fe_species_ger_nfi_2012} species code vector}
\usage{
is_fe_species_ger_nfi_2012(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the
\code{fe_species_ger_nfi_2012} class
}
\description{
Check if an Object is a \strong{fe_species_ger_nfi_2012} species code vector
}
\examples{
spec_ids <- new_fe_species_ger_nfi_2012(
  as.character(c(50, 50, 20, 20, 100, 100, 100))
)
is_fe_species_ger_nfi_2012(spec_ids)

}
