% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_species_master.R
\name{is_fe_species_master}
\alias{is_fe_species_master}
\title{Check if an Object is a \strong{fe_species_master} species code vector}
\usage{
is_fe_species_master(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the
\code{fe_species_master} class
}
\description{
Check if an Object is a \strong{fe_species_master} species code vector
}
\examples{
spec_ids <- new_fe_species_master(
  c("picea_001", "fagus_001", "quercus_002", "quercus_001")
)
is_fe_species_master(spec_ids)

}
