% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_species_tum_wwk_long.R
\name{is_fe_species_tum_wwk_long}
\alias{is_fe_species_tum_wwk_long}
\title{Check if an Object is a \strong{fe_species_tum_wwk_long} species code vector}
\usage{
is_fe_species_tum_wwk_long(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the
\code{fe_species_tum_wwk_long} class
}
\description{
Check if an Object is a \strong{fe_species_tum_wwk_long} species code vector
}
\examples{
spec_ids <- new_fe_species_tum_wwk_long(
  as.character(c(70, 61, 88, 88, 10, 971, 32))
)
is_fe_species_tum_wwk_long(spec_ids)

}
