% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_stand.R
\name{is_fe_stand}
\alias{is_fe_stand}
\title{Check if an Object is an \strong{fe_stand}}
\usage{
is_fe_stand(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{fe_stand} class
}
\description{
Check if an Object is an \strong{fe_stand}
}
\examples{
strange <- "xyzabc"
is_fe_stand(strange)

is_fe_stand(norway_spruce_1_fe_stand)

}
