% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_tally_list.R
\name{is_fe_tally_list}
\alias{is_fe_tally_list}
\title{Check if an Object is an \strong{fe_tally_list}}
\usage{
is_fe_tally_list(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{fe_tally_list}
class
}
\description{
Check if an Object is an \strong{fe_tally_list}
}
\keyword{internal}
