% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_stand_helpers_construction.R
\name{make_trees_dataframe}
\alias{make_trees_dataframe}
\title{Build a 'trees' data.frame (tibble) as Required for \emph{fe_stand} and
\emph{fe_stand_spatial} Objects}
\usage{
make_trees_dataframe(
  x,
  tree_id_col,
  species_id_col,
  time_yr_col,
  dbh_cm_col,
  layer_key_col,
  age_yr_col,
  height_m_col,
  crown_base_height_m_col,
  crown_radius_m_col,
  removal_col,
  ingrowth_col,
  n_rep_ha_col
)
}
\arguments{
\item{x}{Input data frame (tibble)}

\item{tree_id_col}{See documentation of \code{\link{fe_stand}}}

\item{species_id_col}{See documentation of \code{\link{fe_stand}}}

\item{time_yr_col}{See documentation of \code{\link{fe_stand}}}

\item{dbh_cm_col}{See documentation of \code{\link{fe_stand}}}

\item{layer_key_col}{See documentation of \code{\link{fe_stand}}}

\item{age_yr_col}{See documentation of \code{\link{fe_stand}}}

\item{height_m_col}{See documentation of \code{\link{fe_stand}}}

\item{crown_base_height_m_col}{See documentation of \code{\link{fe_stand}}}

\item{crown_radius_m_col}{See documentation of \code{\link{fe_stand}}}

\item{removal_col}{See documentation of \code{\link{fe_stand}}}

\item{ingrowth_col}{See documentation of \code{\link{fe_stand}}}

\item{n_rep_ha_col}{See documentation of \code{\link{fe_stand}} and
\code{\link{fe_stand_spatial}}}
}
\value{
A data.frame (tibble) that can be used as the list element 'trees' in
an \emph{fe_stand} or \emph{fe_stand_spatial} object
}
\description{
Build a 'trees' data.frame (tibble) as Required for \emph{fe_stand} and
\emph{fe_stand_spatial} Objects
}
\keyword{internal}
