% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_species_generic_methods.R
\name{new_fe_species}
\alias{new_fe_species}
\title{Generic Constructor for \strong{fe_species} Objects}
\usage{
new_fe_species(x = character(), class_name)
}
\arguments{
\item{x}{A \code{character} vector}

\item{class_name}{Name of the class to be created. Must be one of the
supported names (first column of data \code{species_codings} with the
prefix "fe_species_")}
}
\value{
An object of class \code{class_name} if x is \code{character} and
\code{class_name} is among the supported names. Otherwise, the function
terminates with an error.
}
\description{
Generic Constructor for \strong{fe_species} Objects
}
\keyword{internal}
