% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_species_bavrn_state.R
\name{new_fe_species_bavrn_state}
\alias{new_fe_species_bavrn_state}
\title{Constructor for the \strong{fe_species_bavrn_state} Class}
\usage{
new_fe_species_bavrn_state(x = character())
}
\arguments{
\item{x}{An appropriate \code{character} vector}
}
\value{
An object of class \code{fe_species_bavrn_state}
}
\description{
Should be used by expert users only who know exactly what they are doing.
Other users please take the function \code{\link{fe_species_bavrn_state}}
for creating an object of that class.
}
\examples{
# Constructing a fe_species_bavrn_state object from scratch
# Use fe_species_bavrn_state() if you are not absolutely sure
spec_ids <- new_fe_species_bavrn_state(
  as.character(c(40, 40, 30, 30, 60, 60, 60))
)

}
