% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_species_ger_nfi_2012.R
\name{new_fe_species_ger_nfi_2012}
\alias{new_fe_species_ger_nfi_2012}
\title{Constructor for the \strong{fe_species_ger_nfi_2012} Class}
\usage{
new_fe_species_ger_nfi_2012(x = character())
}
\arguments{
\item{x}{An appropriate \code{character} vector}
}
\value{
An object of class \code{fe_species_ger_nfi_2012}
}
\description{
Should be used by expert users only who know exactly what they are doing.
Other users please take the function \code{\link{fe_species_ger_nfi_2012}}
for creating an object of that class.
}
\examples{
# Constructing a fe_species_ger_nfi_2012 object from scratch
# Use fe_species_ger_nfi_2012() if you are not absolutely sure
spec_ids <- new_fe_species_ger_nfi_2012(
  as.character(c(50, 50, 20, 20, 100, 100, 100))
)

}
