% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_species_master.R
\name{new_fe_species_master}
\alias{new_fe_species_master}
\title{Constructor for the \strong{fe_species_master} Class}
\usage{
new_fe_species_master(x = character())
}
\arguments{
\item{x}{An appropriate \code{character} vector}
}
\value{
An object of class \code{fe_species_master}
}
\description{
Should be used by expert users only who know exactly what they are doing.
Other users please take the function \code{\link{fe_species_master}}
for creating an object of that class.
}
\examples{
# Constructing a fe_species_master object from scratch
# Use fe_species_master() if you are not absolutely sure
spec_ids <- new_fe_species_master(
  c("picea_001", "fagus_001", "quercus_002", "quercus_001")
)

}
