% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_species_tum_wwk_long.R
\name{new_fe_species_tum_wwk_long}
\alias{new_fe_species_tum_wwk_long}
\title{Constructor for the \strong{fe_species_tum_wwk_long} Class}
\usage{
new_fe_species_tum_wwk_long(x = character())
}
\arguments{
\item{x}{An appropriate \code{character} vector}
}
\value{
An object of class \code{fe_species_tum_wwk_long}
}
\description{
Should be used by expert users only who know exactly what they are doing.
Other users please take the function \code{\link{fe_species_tum_wwk_long}}
for creating an object of that class.
}
\examples{
# Constructing a fe_species_tum_wwk_long object from scratch
# Use fe_species_tum_wwk_long() if you are not absolutely sure
spec_ids <- new_fe_species_tum_wwk_long(
  as.character(c(70, 61, 88, 88, 10, 971, 32))
)

}
