% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_tally_list.R
\name{new_fe_tally_list}
\alias{new_fe_tally_list}
\title{Constructor for the \strong{fe_tally_list} Class}
\usage{
new_fe_tally_list(x = list(), ..., class = character())
}
\arguments{
\item{x}{An appropriate \code{list} object}

\item{...}{Additional arguments required for enabling subclasses of
\code{fe_tally_list}}

\item{class}{A Character string required for enabling subclasses of
\code{fe_tally_list}}
}
\value{
An object of class \code{fe_tally_list}
}
\description{
Should be used by expert users only who know exactly what they are doing.
Other users please take the function \code{\link{fe_stand}} for creating an
object of that class.
}
\keyword{internal}
