% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_yield_table.R
\name{new_fe_yield_table}
\alias{new_fe_yield_table}
\title{Constructor for the \strong{fe_yield_table} Class}
\usage{
new_fe_yield_table(x = list())
}
\arguments{
\item{x}{An appropriate \code{list} object}
}
\value{
An object of class \code{fe_yield_table}
}
\description{
Provided for expert users only who know exactly what they are doing. Other
users please take the function \code{\link{fe_yield_table}} for creating an
object of that class.
}
\examples{
  # Some highly motivated object, even a list
  x <- list(my_dream = "I want to be a yield table!")
  x <- new_fe_yield_table(x) # let's help the guy
  is_fe_yield_table(x)       # nice - it worked?!

  # But here's the error. That's why you should not use the bare constructor
  # if you don't know what you are doing.
  try(
    validate_fe_yield_table(x)
  )

}
