% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_ccircle_spatial_notrees_methods.R
\name{plot.fe_ccircle_spatial_notrees}
\alias{plot.fe_ccircle_spatial_notrees}
\title{Plot an fe_ccircle_spatial_notrees Object}
\usage{
\method{plot}{fe_ccircle_spatial_notrees}(x, ...)
}
\arguments{
\item{x}{An \code{\link{fe_ccircle_spatial_notrees}} object}

\item{...}{Additional arguments, not used in
\code{plot.fe_ccircle_spatial_notrees}}
}
\value{
A map (ggplot2) of the plot layout including the trees with
coordinates
}
\description{
For an object of class\code{\link{fe_ccircle_spatial_notrees}} only the
concentric circles are plotted at their correct positions.
}
\examples{
  # Abuse input data, that would actually allow for a full fe_ccircle_spatial
  # object to construct a fe_ccircle_spatial_notrees object.
  x <- spruce_pine_ccircle_raw
  x <- x |> fe_ccircle_spatial_notrees()

  # Plot it
  plot(x)

}
