% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standing_area_gnfi3.R
\name{stndg_area_gnfi3_spec_convert}
\alias{stndg_area_gnfi3_spec_convert}
\title{Helper function for cascaded species conversion when applying the standing
area estimation of the German National forest inventory 2012.
Try to convert given species_id into ger_nfi_2012. If this causes an error
convert to tum_wwk_short instead (to be used in the default version of the
function).}
\usage{
stndg_area_gnfi3_spec_convert(species_id)
}
\description{
Helper function for cascaded species conversion when applying the standing
area estimation of the German National forest inventory 2012.
Try to convert given species_id into ger_nfi_2012. If this causes an error
convert to tum_wwk_short instead (to be used in the default version of the
function).
}
\keyword{internal}
