% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_ccircle_spatial_notrees_methods.R
\name{summary.fe_ccircle_spatial_notrees}
\alias{summary.fe_ccircle_spatial_notrees}
\title{summary For a fe_ccircle_spatial_notrees object}
\usage{
\method{summary}{fe_ccircle_spatial_notrees}(object, ...)
}
\arguments{
\item{object}{An \code{\link{fe_ccircle_spatial_notrees}} object}

\item{...}{Additional arguments, not used in
\code{plot.fe_ccircle_spatial_notrees}}
}
\value{
An empty data frame
}
\description{
For the time being, this is function serves only to obtain a controled
output when \code{\link{summary}} is called for an object of class
\code{\link{fe_ccircle_spatial_notrees}}. This output is just an empty
data data frame.
}
\examples{
  # Abuse input data, that would actually allow for a full fe_ccircle_spatial
  # object to construct a fe_ccircle_spatial_notrees object.
  x <- spruce_pine_ccircle_raw
  x <- x |> fe_ccircle_spatial_notrees()

  # Make the dummy summary
  summary(x)

}
