% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_ccircle_spatial_notrees.R
\name{validate_fe_ccircle_spatial_notrees}
\alias{validate_fe_ccircle_spatial_notrees}
\title{Validate an \strong{fe_ccircle_spatial_notrees} Object}
\usage{
validate_fe_ccircle_spatial_notrees(x)
}
\arguments{
\item{x}{An object that is expected to be a correct
\code{fe_ccircle_spatial_notrees} object}
}
\value{
Returns \code{x}, but this function is mainly called for its side
effect which is pointing out any violations of the
\code{fe_ccircle_spatial_notrees} object specifications. In case of such
violations, the function will terminate with an error.
}
\description{
Regular users will not require this function. Expert users will want to use
it in combination with the constructor
\code{\link{new_fe_ccircle_spatial_notrees}}. Regular users, please construct
\code{fe_ccircle_spatial_notrees} objects with
\code{\link{fe_ccircle_spatial_notrees}}.
}
\examples{
  # Validate the example fe_ccircle_spatial object
  spruce_pine_ccircle_spatial_notrees |>
    validate_fe_ccircle_spatial_notrees()


}
