% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_species_generic_methods.R
\name{validate_fe_species}
\alias{validate_fe_species}
\title{Generic Validator for \strong{fe_species} Objects}
\usage{
validate_fe_species(x)
}
\arguments{
\item{x}{An object that is expected to be a correct instance of one of the
supported \code{fe_species} classes}
}
\value{
Returns \code{x} if it passes the validation, terminates with an
error otherwise
}
\description{
Generic Validator for \strong{fe_species} Objects
}
\keyword{internal}
