% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_stand_spatial.R
\name{validate_fe_stand_spatial}
\alias{validate_fe_stand_spatial}
\title{Validate an \strong{fe_stand_spatial} Object}
\usage{
validate_fe_stand_spatial(x)
}
\arguments{
\item{x}{An object that is expected to be a correct \code{fe_stand_spatial}
object}
}
\value{
Returns \code{x}, but this function is mainly called for its side
effect which is pointing out any violations of the \code{fe_stand_spatial}
object specifications. In case of such violations, the function will
terminate with an error.
}
\description{
Regular users will not require this function. Expert users will want to use
it in combination with the constructor \code{\link{new_fe_stand_spatial}}.
Regular users, please construct \code{fe_stand_spatial} objects with
\code{\link{fe_stand_spatial}}.
}
\examples{
  # Validate the example fe_stand_spatial object
  mm_forest_1_fe_stand_spatial |> validate_fe_stand_spatial()

}
