% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_tally_list.R
\name{validate_fe_tally_list}
\alias{validate_fe_tally_list}
\title{Validate an \strong{fe_tally_list} Object}
\usage{
validate_fe_tally_list(x)
}
\arguments{
\item{x}{An object that is expected to be a correct \code{fe_tally_list}
object}
}
\value{
Returns \code{x}, but this function is mainly called for its side
effect which is pointing out any violations of the \code{fe_tally_list}
object specifications. In case of such violations, the function will
terminate with an error.
}
\description{
Regular users will not require this function. Expert users will want to use
it in combination with the constructor \code{\link{new_fe_tally_list}}.
Regular users, please construct \code{fe_tally_list} objects with
\code{\link{fe_tally_list}}.
}
\keyword{internal}
