% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_yield_table.R
\name{validate_fe_yield_table}
\alias{validate_fe_yield_table}
\title{Validate a Candidate For an fe_yield_table Object}
\usage{
validate_fe_yield_table(x)
}
\arguments{
\item{x}{The candidate object to be validated}
}
\value{
If \code{x} is not a valid \code{fe_yield_table} object, the function
will terminate with an error. Otherwise, \code{x} will be returned.
}
\description{
Validate a Candidate For an fe_yield_table Object
}
\examples{
 validate_fe_yield_table(fe_ytable_beech_wiedemann_moderate_1931)
 validate_fe_yield_table(fe_ytable_larch_schober_moderate_1946)
 validate_fe_yield_table(fe_ytable_pine_wiedemann_moderate_1943)
 validate_fe_yield_table(fe_ytable_spruce_gehrhardt_moderate_1921)
 validate_fe_yield_table(fe_ytable_douglas_schober_moderate_1956)
 validate_fe_yield_table(fe_ytable_spruce_wiedemann_moderate_1936_42)
 validate_fe_yield_table(fe_ytable_silver_fir_hausser_moderate_1956)

}
