% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_species_bavrn_state_short.R
\name{vec_ptype_abbr.fe_species_bavrn_state_short}
\alias{vec_ptype_abbr.fe_species_bavrn_state_short}
\title{Abbreviation for the \emph{fe_species_bavrn_state_short} Type}
\usage{
\method{vec_ptype_abbr}{fe_species_bavrn_state_short}(x, ...)
}
\arguments{
\item{x}{An object of type \code{fe_species_bavrn_state_short}}

\item{...}{Other parameters (not used)}
}
\value{
The abbreviation to be displayed for the species coding
(\code{character}) in tibbles and in \code{str()}
}
\description{
Provide an abbreviated name for the class \code{fe_species_bavrn_state_short}
to be displayed in tibbles and \code{str()}
}
\examples{
spec_ids <- fe_species_bavrn_state_short(as.character(c(2, 4, 6)))
vctrs::vec_ptype_abbr(spec_ids)
str(spec_ids)

}
