% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_species_ger_nfi_2012.R
\name{vec_ptype_abbr.fe_species_ger_nfi_2012}
\alias{vec_ptype_abbr.fe_species_ger_nfi_2012}
\title{Abbreviation for the \emph{fe_species_ger_nfi_2012} Type}
\usage{
\method{vec_ptype_abbr}{fe_species_ger_nfi_2012}(x, ...)
}
\arguments{
\item{x}{An object of type \code{fe_species_ger_nfi_2012}}

\item{...}{Other parameters (not used)}
}
\value{
The abbreviation to be displayed for the species coding
(\code{character}) in tibbles and in \code{str()}
}
\description{
Provide an abbreviated name for the class \code{fe_species_ger_nfi_2012} to
be displayed in tibbles and \code{str()}
}
\examples{
spec_ids <- fe_species_ger_nfi_2012(as.character(c(10, 20, 50)))
vctrs::vec_ptype_abbr(spec_ids)
str(spec_ids)

}
