% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_species_tum_wwk_long.R
\name{vec_ptype_abbr.fe_species_tum_wwk_long}
\alias{vec_ptype_abbr.fe_species_tum_wwk_long}
\title{Abbreviation for the \emph{fe_species_tum_wwk_long} Type}
\usage{
\method{vec_ptype_abbr}{fe_species_tum_wwk_long}(x, ...)
}
\arguments{
\item{x}{An object of type \code{fe_species_tum_wwk_long}}

\item{...}{Other parameters (not used)}
}
\value{
The abbreviation to be displayed for the species coding
(\code{character}) in tibbles and in \code{str()}
}
\description{
Provide an abbreviated name for the class \code{fe_species_tum_wwk_long} to
be displayed in tibbles and \code{str()}
}
\examples{
spec_ids <- fe_species_tum_wwk_long(as.character(c(10, 50, 87, 813)))
vctrs::vec_ptype_abbr(spec_ids)
str(spec_ids)

}
