% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wood_density_tum_wwk_short.R
\docType{data}
\name{wood_density_tum_wwk_short}
\alias{wood_density_tum_wwk_short}
\title{Basic Wood Density Values for the \emph{tum_wwk_short} Species Coding}
\format{
A tibble with one row per species of the \emph{tum_wwk_short} coding and
the columns:
\describe{
\item{species_id}{the species code, an object of class
\code{fe_species_tum_wwk_short}}
\item{wood_density}{basic wood density in \eqn{t/m^3} (i.e. oven-dry mass
per unit of green volume)}
}
}
\usage{
wood_density_tum_wwk_short
}
\description{
Basic wood density (oven-dry mass over green volume) for each
species of the \emph{tum_wwk_short} species coding (see
\code{\link{species_codings}}). These values are used for converting wood
volumes into masses, e.g. when calculating species shares by biomass.
}
\examples{
wood_density_tum_wwk_short

}
\keyword{datasets}
